Refs #11688; Add progress bar to NetexFull exports
