+++ app/models/export/netex_full.rb
@@ -21,4 +21,12 @@ class Export::NetexFull < Export::Base
     build_netex_document
     document.temp_file
   end
+
+  def operation_progress_weight(operation_name)
+    operation_name.to_sym == :site_frame ? 90 : 10.0/4
+  end
+
+  def operations_progress_total_weight
+    100
+  end
 end
